+++ Krowi_AchievementFilter/Krowi_AchievementFilter.lua
@@ -11,8 +11,9 @@ addon.Util = LibStub("Krowi_Util-1.0");
 
 -- [[ Binding names ]] --
 BINDING_HEADER_AF_NAME = addon.MetaData.Title;
-BINDING_NAME_AF_OPEN_TAB1 = addon.L["Toggle Achievement Window to "] .. addon.L["Expansions"];
-BINDING_NAME_AF_OPEN_TAB2 = addon.L["Toggle Achievement Window to "] .. addon.L["Events"];
+BINDING_NAME_AF_OPEN_TAB1 = addon.L["Toggle "] .. addon.L["Expansions"] .. addon.L[" tab"];
+BINDING_NAME_AF_OPEN_TAB2 = addon.L["Toggle "] .. addon.L["Events"] .. addon.L[" tab"];
+BINDING_NAME_AF_OPEN_TAB3 = addon.L["Toggle "] .. addon.L["Player vs. Player"] .. addon.L[" tab"];
 
 -- [[ Guild view ]] --
 function addon.InGuildView()